Disable soft-limit on 4store!

+++ scripts/4store.sh
@@ -1,10 +1,13 @@
 #! /bin/sh
 
-killall 4s-httpd
-killall 4s-backend
-
 dbname=biorubyrdftest
-4s-backend-setup $dbname
+
+if [ "$1" = "-r" ]; then
+  killall 4s-httpd
+  killall 4s-backend
+  4s-backend-setup $dbname
+fi
+
 4s-backend $dbname
 4s-httpd -p 8000 $dbname
